requirements.txt
base58==2.1.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
typing_extensions==4.12.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
